Refactor(#14) : swagger 수정, 패키지 이름 수정
