+++ src/main/java/io/urdego/urdego_game_service/controller/client/content/ContentServiceClient.java
@@ -1,6 +1,6 @@
-package io.urdego.urdego_game_service.common.client.content;
+package io.urdego.urdego_game_service.controller.client.content;
 
-import io.urdego.urdego_game_service.common.client.content.dto.ContentRes;
+import io.urdego.urdego_game_service.controller.client.content.dto.ContentRes;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
