+++ src/main/java/io/urdego/urdego_game_service/controller/client/content/dto/ContentRes.java
@@ -1,4 +1,4 @@
-package io.urdego.urdego_game_service.common.client.content.dto;
+package io.urdego.urdego_game_service.controller.client.content.dto;
 
 public record ContentRes(
         Long contentId,
