+++ src/main/java/io/urdego/urdego_game_service/domain/round/service/RoundServiceImpl.java
@@ -4,8 +4,8 @@ import io.urdego.urdego_game_service.controller.round.dto.request.AnswerReq;
 import io.urdego.urdego_game_service.controller.round.dto.request.QuestionReq;
 import io.urdego.urdego_game_service.controller.round.dto.response.AnswerRes;
 import io.urdego.urdego_game_service.controller.round.dto.response.QuestionRes;
-import io.urdego.urdego_game_service.common.client.content.ContentServiceClient;
-import io.urdego.urdego_game_service.common.client.content.dto.ContentRes;
+import io.urdego.urdego_game_service.controller.client.content.ContentServiceClient;
+import io.urdego.urdego_game_service.controller.client.content.dto.ContentRes;
 import io.urdego.urdego_game_service.common.exception.ExceptionMessage;
 import io.urdego.urdego_game_service.common.exception.round.QuestionException;
 import io.urdego.urdego_game_service.domain.room.entity.Room;
